+++ src/BrandNode.h
@@ -5,22 +5,22 @@
 #include <Geode/Geode.hpp>
 
 namespace cw::brand {
-    class BrandingNode final : public cocos2d::CCNode {
+    class BrandNode final : public cocos2d::CCNode {
     private:
         struct Impl;
         std::unique_ptr<Impl> m_impl;
 
     protected:
-        BrandingNode();
-        ~BrandingNode();
+        BrandNode();
+        ~BrandNode();
 
         void retryRemoteLoad(geode::LazySprite* sender);
         void cancelRemoteLoad(CCNode* sender);
 
-        bool init(geode::MDTextArea* container, std::string dev, geode::ZStringView modId);
+        bool init(geode::NineSlice* container, std::string dev, geode::ZStringView modId);
 
     public:
-        static BrandingNode* create(geode::MDTextArea* container, std::string dev, geode::ZStringView modId = GEODE_MOD_ID);
+        static BrandNode* create(geode::NineSlice* container, std::string dev, geode::ZStringView modId = GEODE_MOD_ID);
 
         void loadBrand();