+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "BrandingNode.h"
+#include "BrandNode.h"
 
 #include <Geode/Geode.hpp>
 
@@ -13,40 +13,59 @@ static const ZStringView urlGeode = "https://geode-sdk.org/mods/";
 
 class $nodeModify(BrandingModPopup, ModPopup) {
     struct Fields final {
+        BrandNode* branding = nullptr;
         Ref<MDTextArea> textArea = nullptr;
-        BrandingNode* branding = nullptr;
 
         float height = 0.f;
     };
 
+    // nesting hell
     void modify() {
-        if (auto md = typeinfo_cast<MDTextArea*>(getChildByIDRecursive("textarea"))) {
-            auto username = getGitUsername();
-            auto const mod = getModID();
+        if (auto self = reinterpret_cast<FLAlertLayer*>(this)) {
+            log::trace("Modifying mod popup for branding");
 
-            log::debug("Creating brand node for {} by {}", mod, username);
+            // node reordering lol
+            if (auto displayNode = self->m_mainLayer->getChildByType<CCNode*>(2)) {
+                log::trace("Found popup display node");
 
-            if (auto brand = BrandingNode::create(md, std::move(username), mod)) {
-                auto f = m_fields.self();
+                if (auto md = typeinfo_cast<MDTextArea*>(displayNode->querySelector("right-column > description-container > textarea"))) {
+                    log::trace("Found mod description markdown container");
 
-                f->textArea = md;
-                f->height = md->getScaledContentHeight();
+                    if (auto container = md->getChildByType<NineSlice*>()) {
+                        auto username = getGitUsername();
+                        auto const mod = getModID();
 
-                brand->setPositionX(brand->getPositionX() + 7.794f);
-                brand->setZOrder(-9);
+                        log::debug("Creating brand node for {} by {}", mod, username);
 
-                f->branding = brand;
+                        if (auto brand = BrandNode::create(container, std::move(username), mod)) {
+                            auto f = m_fields.self();
 
-                md->addChild(f->branding);
+                            f->textArea = md;
+                            f->height = md->getScaledContentHeight();
 
-                log::info("Added brand node for {} by {}", mod, brand->getDeveloper());
+                            brand->setZOrder(-9);
+
+                            f->branding = brand;
+
+                            md->addChildAtPosition(f->branding, Anchor::BottomRight, {(container->getScaledContentWidth() - md->getScaledContentWidth()) * 0.5f, 0.f});
+
+                            log::info("Added brand node for {} by {}", mod, brand->getDeveloper());
+                        } else {
+                            log::error("Couldn't create brand node");
+                        };
+
+                        schedule(schedule_selector(BrandingModPopup::updateBrandSize));
+                    } else {
+                        log::error("Couldn't find mod description container background");
+                    };
+                } else {
+                    log::error("Couldn't find mod description markdown container");
+                };
             } else {
-                log::error("Couldn't create brand node");
+                log::error("Couldn't find mod popup display node");
             };
-
-            schedule(schedule_selector(BrandingModPopup::updateBrandSize));
         } else {
-            log::error("Couldn't find mod description container");
+            log::error("Couldn't cast to FLAlertLayer");
         };
     };